+++ test/helper.rb
@@ -7,6 +7,6 @@ require 'shoulda-context'
 $LOAD_PATH.unshift(File.join(File.dirname(__FILE__), '..', 'lib'))
 $LOAD_PATH.unshift(File.dirname(__FILE__))
 
-require 'taglib'
+require 'taglib_plus'
 
 HAVE_ENCODING = !RUBY_VERSION.start_with?('1.8')